refactor: 인증 코드 저장 로직 추상화
EmailAuthService가 AuthCodeStorage 인터페이를 의존하도록 만들어서 추후에 Redis가 아닌 다른 저장소를 사용하더라도 코드 변경 없이 동작할 수 있도록 설계
