+++ back/src/main/java/com/ll/goohaeyou/jobPost/jobPost/application/EssentialService.java
@@ -1,12 +1,11 @@
 package com.ll.goohaeyou.jobPost.jobPost.application;
 
 import com.ll.goohaeyou.global.exception.EntityNotFoundException;
-import com.ll.goohaeyou.jobPost.jobPost.application.dto.JobPostDto;
-import com.ll.goohaeyou.jobPost.jobPost.application.dto.WriteJobPostRequest;
 import com.ll.goohaeyou.jobPost.jobPost.domain.Essential;
 import com.ll.goohaeyou.jobPost.jobPost.domain.JobPostDetail;
 import com.ll.goohaeyou.jobPost.jobPost.domain.repository.EssentialRepository;
 import com.ll.goohaeyou.jobPost.jobPost.domain.repository.JobPostDetailRepository;
+import com.ll.goohaeyou.member.member.domain.type.Gender;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,12 +18,11 @@ public class EssentialService {
     private final JobPostDetailRepository jobPostDetailRepository;
 
     @Transactional
-    public void create(JobPostDto jobPostDto, WriteJobPostRequest request) {
-        JobPostDetail jobPostDetail = jobPostDetailRepository.findById(jobPostDto.getId())
+    public void create(Long jobPostId, int minAge, Gender gender) {
+        JobPostDetail jobPostDetail = jobPostDetailRepository.findById(jobPostId)
                 .orElseThrow(EntityNotFoundException.PostNotExistsException::new);
 
-        Essential newEssential = Essential.create(request.minAge(),
-                request.gender(), jobPostDetail);
+        Essential newEssential = Essential.create(minAge, gender, jobPostDetail);
 
         essentialRepository.save(newEssential);
     }
